Fix pseudo console bug and RPLA checkbox label
The pseudo console would not be inserted since body.appendChild() was
called before body existed. Now it will be inserted at the end, since
it does not affect page jumping and since it is a dev tool anyway.

The checkbox for remove page link anchors had a <pre> element in it,
which created unwanted line breaks. It is now removed.